backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
email-validator
# Settings
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client (for Kakao OAuth)
httpx==0.26.0

# Development
python-dotenv==1.0.0
